Fix serial Write and Read

+++ src/stage2/Drivers/Serial.hpp
@@ -8,7 +8,7 @@ class Serial
         static void Initialize();
 
         static void WriteByte(uint8_t byte);
-        static int16_t ReadByte(uint8_t byte);
+        static uint8_t ReadByte(uint8_t byte);
 
     private:
         static constexpr const uint32_t SERIAL_PORT1 = 0x3f8;
